Make sure the user really is logged in before welcoming them
Remove this.thisUser which suffers from exactly the same persistence in memory in an Activity as both this.patientData and this.shouldShowCountryPicker. The correct test for whether a user is logged in or not is this.userService.hasUser, which is the only value forcibly reset when you log out.

This fixes the remaining hole I’ve been hitting (on staging) as a logged-out US user in a UK phone environment, where I was being redirected via the shouldShowCountryPicker instead of being asked to log in.